# docs/swagger.yaml
components:
  schemas:
...
    models.ErrorResponse:
      properties:
        error:
          example: Message d'erreur
          type: string
        message:
          example: Détail de l'erreur
          type: string
      type: object
...
  securitySchemes:
    bearerauth:
      bearerFormat: JWT
      scheme: bearer
      type: http
...
info:
  contact:
    name: Équipe Transat
  description: API Backend pour l'application Transat - l'application du campus de
    Nantes d'IMT Atlantique
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: API Transat
  version: "2.0"
openapi: 3.1.0
paths:
...
  /statistics/endpoints:
    get:
      description: Récupère les statistiques d'utilisation détaillées pour tous les
        endpoints de l'API
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Statistiques par endpoint récupérées avec succès
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Non autorisé
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Erreur serveur
      security:
      - BearerAuth: []
      summary: Obtenir les statistiques par endpoint
      tags:
      - Statistics
  /statistics/global:
    get:
      description: Récupère les statistiques globales d'utilisation de l'API (tous
        endpoints confondus)
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Statistiques globales récupérées avec succès
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Non autorisé
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Erreur serveur
      security:
      - BearerAuth: []
      summary: Obtenir les statistiques globales
      tags:
      - Statistics
...
servers:
- description: Localhost
  url: http://localhost:3000/api
- description: Production
  url: https://api.transat.dev/api
...